\name{cleanup.gds}
\alias{cleanup.gds}
\title{Clean up fragments}
\description{
    Clean up the fragments of a CoreArray Genomic Data Structure (GDS) file.
}

\usage{
cleanup.gds(filename, sort=TRUE, verbose=TRUE)
}
\arguments{
    \item{filename}{the file name of a GDS file to be opened}
    \item{sort}{if \code{TRUE}, rearrange data blocks when defragmenting:
        all node header blocks are placed first in depth-first traversal
        order with folder headers prioritized before other node headers,
        followed by the data blocks grouped by the node that owns them. The
        groups are ordered by the total size of all the streams a node keeps
        (small to large), and within a node the blocks are sorted by size
        (small to large) then by ID. This clusters the GDS tree structure
        metadata at the front of the file, maximizing page cache hit rate
        during tree navigation, and keeps each node's streams adjacent}
    \item{verbose}{if \code{TRUE}, show information}
}
\value{
    None.
}

\author{Xiuwen Zheng}
\seealso{
    \code{\link{openfn.gds}}, \code{\link{createfn.gds}},
    \code{\link{closefn.gds}}
}

\examples{
# cteate a GDS file
f <- createfn.gds("test.gds")

# commom types
add.gdsn(f, "int", val=1:10000)
L <- -2500:2499
add.gdsn(f, "int.matrix", val=matrix(L, nrow=100, ncol=50))

# save a .RData object
obj <- list(X=1:10, Y=seq(1, 10, 0.1))
save(obj, file="tmp.RData")
addfile.gdsn(f, "tmp.RData", filename="tmp.RData")

f

# close the GDS file
closefn.gds(f)


# clean up fragments
cleanup.gds("test.gds")


# open ...
(f <- openfn.gds("test.gds"))
closefn.gds(f)


# delete the temporary file
unlink("test.gds", force=TRUE)
}

\keyword{GDS}
\keyword{utilities}
