\name{gdsCFuncList}
\alias{gdsCFuncList}
\title{C functions}
\description{
	A list of C functions in the package.
}

\usage{
gdsCFuncList()
}
\value{
	A list of C functions registered via "R_RegisterCCallable" in the gdsfmt
package.
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng}

\examples{
gdsCFuncList()
}

\keyword{GDS}
\keyword{utilities}
