\name{gdsn}
\alias{gdsn}
\title{the class of a node of GDS file}
\description{
	the class of the node of hierarchical structure from a GDS file, and its instance is
returned from \code{\link[gdsfmt]{add.gdsn}} or \code{\link[gdsfmt]{index.gdsn}}.
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}
\seealso{\code{\link[gdsfmt]{createfn.gds}}, \code{\link[gdsfmt]{closefn.gds}}}

\keyword{utilities}
