\name{gdsUnitTest}
\alias{gdsUnitTest}
\title{Unit testing}
\description{
	Run the test R codes in the folder of 'unitTests'
}

\usage{
gdsUnitTest()
}

\value{
	None.
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}

\keyword{GDS}
\keyword{utilities}
