\name{print.gdsclass}
\alias{print.gdsclass}
\title{Show the information of class "gdsclass"}
\description{
	Displays an object of class "gdsclass", in a CoreArray Genomic Data Structure (GDS) file.
}
\usage{
\method{print}{gdsclass}(x, ...)
}
\arguments{
	\item{x}{an object of class \code{\link[gdsfmt]{gdsclass}}, a CoreArray GDS file}
	\item{...}{the arguments passed to or from other methods}
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}
\seealso{\code{\link[gdsfmt]{print.gdsn}}}

\keyword{utilities}
