\name{gdsn.class}
\docType{class}
\alias{gdsn.class}

\title{the class of variable node in the GDS file}
\description{
	The class of variable node in the GDS file.
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng}
\seealso{
	\code{\link{add.gdsn}}, \code{\link{read.gdsn}},
	\code{\link{write.gdsn}}
}

\keyword{GDS}
\keyword{utilities}
