\name{getfile.gdsn}
\alias{getfile.gdsn}
\title{Output a file from a stream container}
\description{
	Get a file from a GDS node of stream container.
}

\usage{
getfile.gdsn(node, out.filename)
}
\arguments{
	\item{node}{an object of class \code{\link{gdsn.class}}, a GDS node}
	\item{out.filename}{the file name of output stream}
}
\value{
	None.
}

\references{\url{http://github.com/zhengxwen/gdsfmt}}
\author{Xiuwen Zheng}
\seealso{
	\code{\link{addfile.gdsn}}
}

\keyword{GDS}
\keyword{utilities}
