\name{lasterr.gds}
\alias{lasterr.gds}
\title{Return the last error}
\description{
	Returns the last error of a CoreArray GDS file.
}

\usage{
  lasterr.gds()
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}

\keyword{utilities}
