\name{lasterr.gds}
\alias{lasterr.gds}
\title{Return the last error message}
\description{
	Get the last error message and clear the error message(s) in the
gdsfmt package.
}

\usage{
lasterr.gds()
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng}

\keyword{GDS}
\keyword{utilities}
