\name{print.gdsclass}
\alias{print.gdsclass}
\title{Print class "gdsclass"}
\description{
    Print an object of class "gdsclass", in a CoreArray Generic Data Structures
  (GDS) file.
}
\usage{
	\method{print}{gdsclass}(x, ...)
}
\arguments{
	\item{x}{An object of class "gdsclass", the CoreArray GDS file.}
	\item{...}{Arguments passed to or from other methods.}
}

\references{\url{http://sourceforge.net/projects/corearray/}}
\author{Xiuwen Zheng \email{zhengx@u.washington.edu}}
\seealso{\code{\link[gdsfmt]{print.gdsn}}}

\keyword{utilities}
