\name{system.gds}
\alias{system.gds}
\title{Get the parameters used in the GDS system}
\description{
	Get a list of parameters used in the GDS system
}

\usage{
system.gds()
}
\value{
	A list.
}

\references{\url{http://github.com/zhengxwen/gdsfmt}}
\author{Xiuwen Zheng}
\examples{
system.gds()
}

\keyword{GDS}
\keyword{utilities}
